improves the reciving mesage logic, send message function and added function deleteMessageFromList

+++ client/src/stores/message.ts
@@ -3,11 +3,20 @@ import { ref } from "vue";
 import { indexedDbService } from "@/services/indexDbServices";
 import type { Message } from "@/types/Message";
 import type { Conversation } from "@/types/Conversation";
-import type { User } from "@/types/User";
+import { useUserStore } from "./user";
 
 export const useMessageStore = defineStore("message", () => {
 	const socket = new WebSocket("ws://localhost:8000/chat/socket");
-	const conversations = ref<{ [key: string]: Message[] }>({});
+
+	const conversations = ref<{
+		[key: string]: {
+			messages: Message[];
+			participant: string;
+			lastMessageDate: string;
+			isActive: boolean;
+		};
+	}>({});
+	const userStore = useUserStore();
 
 	//data for current going conversation
 	const currentConversation = ref<{
@@ -18,8 +27,8 @@ export const useMessageStore = defineStore("message", () => {
 	//handle receiving message
 	socket.onmessage = async (event) => {
 		const msg = JSON.parse(event.data);
-		console.log(msg);
-		// Check for required fields in `msg` before proceeding
+
+		// Validate required field
 		if (
 			!msg.id ||
 			!msg.conversation_id ||
@@ -43,7 +52,7 @@ export const useMessageStore = defineStore("message", () => {
 			status: msg.status,
 		};
 
-		//TODO: first check the conversation in conversations
+		//check for conversation in indesedDb
 		const conversation: Conversation = (await indexedDbService.getRecord(
 			"conversation",
 			message.conversationId
@@ -54,51 +63,59 @@ export const useMessageStore = defineStore("message", () => {
 			const newConversation: Conversation = {
 				id: message.conversationId,
 				participant: message.senderId,
-				unseenMessage_ids: [],
+				unseenMessageIds: [message.id as string],
 				startDate: null,
 				lastMessageDate: message.sendingTime,
 			};
 
+			//TODO: a function to retrive all info about the conversation and update the indesed db
+
+			conversations.value[message.conversationId as string] = {
+				isActive: true,
+				messages: [],
+				lastMessageDate: message.sendingTime as string,
+				participant: message.senderId as string,
+			};
+
 			await indexedDbService.addRecord("conversation", newConversation);
-
-			conversations.value[message.conversationId] = [];
-
-			//the message if of current conv
+		} else {
+			// For recent conversation
 			if (
-				currentConversation.value &&
-				message.senderId == currentConversation.value.recieverId
+				currentConversation.value == null ||
+				currentConversation.value.convId != message.conversationId
 			) {
-				currentConversation.value.convId = message.conversationId;
+				conversation.unseenMessageIds!.push(message.id as string);
 			}
-		}
+			// Ongoing conversation
+			else {
+				// message.status = "read";
+				//TODO: A function to send server this information
+			}
 
-		if (
-			currentConversation.value == null ||
-			currentConversation.value.convId != message.conversationId
-		) {
-			// this in another conversations
-			// retrive the conversation, add the unseen message id and put it back
-			const conversation: Conversation =
-				(await indexedDbService.getRecord(
-					"conversation",
-					message.conversationId
-				)) as Conversation;
+			// Update the last conversation datetime
+			conversations.value[message.conversationId!].lastMessageDate =
+				message.sendingTime as string;
 
-			conversation.unseenMessage_ids.push(message.id);
+			conversation.lastMessageDate = message.sendingTime;
 			await indexedDbService.updateRecord("conversation", conversation);
 		}
 
-		conversations.value[message.conversationId].push(message);
+		// Remove temp messagee(only for the sender)
+		if (message.senderId == userStore.user.id) {
+			if (msg.temp_id) {
+				await indexedDbService.deleteRecord("message", msg.temp_id);
+				deleteMessageFromList(message.conversationId!, msg.temp_id);
+			}
+		}
 
-		await indexedDbService.deleteRecord("message", msg.temp_id);
+		// Append the new message to the current conversation
+		conversations.value[message.conversationId!].messages.push(message);
+
+		// Store the message in IndexedDB
 		await indexedDbService.addRecord("message", message);
 	};
 
-	async function sendMessage(
-		message: string
-		// conversationId: string | null = null,
-		// recieverId: string | null = null
-	) {
+	async function sendMessage(message: string) {
 		if (!currentConversation.value) {
 			return;
 		}
@@ -106,7 +123,6 @@ export const useMessageStore = defineStore("message", () => {
 			currentConversation.value.convId == null &&
 			currentConversation.value.recieverId == null
 		) {
-			console.log(currentConversation.value.recieverId);
 			console.error("Both conversationId and recieverId are null");
 			return;
 		}
@@ -118,17 +134,59 @@ export const useMessageStore = defineStore("message", () => {
 			temp_id: crypto.randomUUID(),
 		};
 
-		const idbData = {
-			id: msg.temp_id,
-			message: msg.message,
+		// Make the data ready for indexedDb and add it
+		const iDbMessage: Message = {
+			id: msg.temp_id.toString(),
+			senderId: userStore.user.id,
 			recieverId: msg.reciever_id,
 			conversationId: msg.conversation_id,
+			message: msg.message,
+			sendingTime: new Date().toISOString(),
+			status: "pending",
 		};
+		await indexedDbService.addRecord("message", iDbMessage);
 
-		const idbResult = await indexedDbService.addRecord("message", idbData);
-		console.log(idbResult);
+		if (!currentConversation.value.convId) {
+			//New conversation
+			conversations.value[msg.temp_id] = {
+				messages: [],
+				isActive: true,
+				participant: msg.reciever_id as string,
+				lastMessageDate: new Date().toISOString(),
+			};
+			conversations.value[msg.temp_id].messages.push(iDbMessage);
+		} else {
+			// Old conversation
+			conversations.value[
+				currentConversation.value.convId
+			]?.messages.push(iDbMessage);
+		}
 
 		socket.send(JSON.stringify(msg));
+	}
+
+	function deleteMessageFromList(
+		convId: string,
+		targetedId: string,
+		candidateIndex = -1
+	) {
+		// Set candidateIndex to the last element on the first call
+		if (candidateIndex === -1) {
+			candidateIndex = conversations.value[convId].messages.length - 1;
+		}
+
+		// Base case: stop if candidateIndex is less than 0
+		if (candidateIndex < 0) {
+			return;
+		}
+		if (
+			conversations.value[convId].messages[candidateIndex].id ==
+			targetedId
+		) {
+			conversations.value[convId].messages.splice(candidateIndex, 1);
+			return;
+		}
+		return deleteMessageFromList(convId, targetedId, candidateIndex - 1);
 	}
 
 	return { currentConversation, conversations, sendMessage };
